fix(util): managed manifest mode validation
